Move discord commands to using shared code (old commands with -o prefix); add location support
