+++ src/lib/discord/commando/commands/starto.js
@@ -1,0 +1,36 @@
+exports.run = async (client, msg, command) => {
+	// const typeArray = Object.keys(client.utilData.types).map(o => o.toLowerCase())
+	let target = { id: msg.author.id, name: msg.author.tag, webhook: false }
+
+	try {
+		// Check target
+		if (!client.config.discord.admins.includes(msg.author.id) && msg.channel.type === 'text') {
+			return await msg.author.send(client.translator.translate('Please run commands in Direct Messages'))
+		}
+		let webhookName
+		const webhookArray = command.find((args) => args.find((arg) => arg.match(client.re.nameRe)))
+		if (webhookArray) webhookName = webhookArray.find((arg) => arg.match(client.re.nameRe))
+		if (webhookName) webhookName = webhookName.replace(client.translator.translate('name'), '')
+		if (client.config.discord.admins.includes(msg.author.id) && msg.channel.type === 'text') target = { id: msg.channel.id, name: msg.channel.name, webhook: false }
+		if (client.config.discord.admins.includes(msg.author.id) && webhookName) target = { name: webhookName.replace(client.translator.translate('name'), ''), webhook: true }
+		const isRegistered = target.webhook
+			? await client.query.selectOneQuery('humans', { name: target.name, type: 'webhook' })
+			: await client.query.countQuery('humans', { id: target.id })
+
+		if (!isRegistered && client.config.discord.admins.includes(msg.author.id) && target.webhook) {
+			return await msg.reply(`Webhook ${target.name} ${client.translator.translate('does not seem to be registered. add it with')} ${client.config.discord.prefix}${client.translator.translate('webhook')} ${client.translator.translate('add')} ${client.translator.translate('<Your-Webhook-url>')}`)
+		}
+		if (!isRegistered && client.config.discord.admins.includes(msg.author.id) && msg.channel.type === 'text') {
+			return await msg.reply(`${msg.channel.name} ${client.translator.translate('does not seem to be registered. add it with')} ${client.config.discord.prefix}${client.translator.translate('channel')} ${client.translator.translate('add')}`)
+		}
+		if (!isRegistered && msg.channel.type === 'dm') {
+			return await msg.react(client.translator.translate('🙅'))
+		}
+		if (target.webhook) target.id = isRegistered.id
+
+		await client.query.updateQuery('humans', { enabled: 1 }, { id: target.id })
+		await msg.react('✅')
+	} catch (err) {
+		client.log.error('start command unhappy:', err)
+	}
+}
